Refactoring math folder and light_and_color.
