+++ scenes/8_shadows.rt
@@ -1,6 +1,6 @@
 A 0.3   255,255,255
 
-L 0,6,-14  0.5  255,255,255
+L 0,6,-14  0.8  255,255,255
 C 10,10,-10 -1,-1,1 70
 
 sp 0,0,0  2  255,139,50
