--- src/control.md
@@ -1,14 +1,0 @@
-
-# Control Key Bindings
-
-This document provides an overview of the key bindings used in the code.
-
-- **Escape Key (ESC):** Closes the window.
-- **Arrow Keys (Up, Down, Left, Right):** Translate the camera along the X and Y axes.
-- **W, S, A, D Keys:** Move the camera along the Z-axis and rotate it around the Y-axis.
-- **Numerical Keypad Keys (KP 4, 5, 6, 7, 8, 9, 0, 1):** Translate selected shapes along the X, Y, and Z axes. Translate selected shapes (additional transformation on the selected object).
-- **KP Add ( + ) and Subtract ( - ):** Scale the selected shape by 2x or 0.5x (if constraints are met).
-- **Right and Left Shift and Control Keys:** Adjust light brightness.
-- **Home and End Keys:** Move the light in the Y direction (up or down).
-- **Page Down and Delete Keys:** Move the light in the X direction (left or right).
-- **Left Shift and Left Control Keys:** Move the light in the Z direction (forward or backward).
